+++ aemedge/scripts/legacy-api.js
@@ -1,4 +1,4 @@
-import { urlByEnvType } from './utils.js';
+import { urlByEnvType } from './utils/index.js';
 
 export function getLegacyAlerts() {
   return `${urlByEnvType()}/content/cmegroup/en/misc/api/content-feeds-for-google-docs/full-alerts-list/jcr:content/main-content-section/section/section-elements/search_sort_filter_d.ssfajax.0.json`;
