+++ aemedge/scripts/utils/index.js
@@ -1,6 +1,7 @@
 export * from './cookie.js';
 export * from './dataCache.js';
 export * from './dom.js';
+export * from './env.js';
 export * from './fetch.js';
 export * from './gtm.js';
 export * from './misc.js';
